Refactor cluster update method
Without this change, the updateUpstreamClusterState method for the
controller handler is long and unwieldy because it needs to handle
several different update requests in slightly different ways. This
change creates a new internal package "gke" and creates new update
functions for each grouping of attributes that can be updated together.
This makes it so that the controller update method need only be
concerned with initiating requests in the right order and enqueuing new
updates on the handler and setting the status on the downstream
resource. It also removes oneoff functions that were used to make
GKE API requests from the "util" package so that GKE API requests are
more neatly handled by the "gke" package.