Read cloud credential from Secret
A Secret representing the service account credential will be created by
Rancher using the CloudCredentials controller. This change adds the
ability to retrieve the credential from a Secret, identified by
`<namespace>:<name>`, where the JSON-encoded credential will be stored
under key googlecredentialConfig-authEncodedJson.

Minor refactoring is done so that the controller handler is in charge of
looking up the credential secret from its own secretCache attribute, and
GetGKEClient and GetServiceClient are condensed to a single function
GetGKEClient that accepts that credential string.