Change GKEClusterNetworkConfig field
The GKEClusterNetworkConfig field of the CRD contained attributes for
setting Network and Subnetwork and was meant to represent the
NetworkConfig field in a GKE Cluster object[1]. However, those fields in
the Go SDK are actually read-only, and the correct way to set the
Network and Subnetwork fields on the Cluster object directly.

This could be done with an internal change in the
newClusterCreateRequest method, but changing the attribute structure of
the CRD to reflect the structure of the GKE object makes the translation
a litte easier to follow.

[1] https://pkg.go.dev/google.golang.org/api/container/v1#NetworkConfig