Rename NodeConfig struct
Rancher's machine driver creates a dynamic schema with name "nodeconfig"
that collides with the static schema generated for the GKE operator's
NodeConfig struct. The result is that repeated calls to
/v3/schemas/nodeConfig may alternately return either the NodeConfig
schema from this operator or a different NodeConfig schema containing
the googleConfig schema for GCP cloud credentials. Rename the parameter
here to avoid this collision.